Code update: minor changes

+++ src/main/java/com/backend/liaison_system/users/student/assumption_of_duty/requests/CreateNewAssumptionOfDuty.java
@@ -18,6 +18,6 @@ public record CreateNewAssumptionOfDuty(
 
         String letterTo,
 
-        int companyLongitude,
-        int companyLatitude
+        double companyLongitude,
+        double companyLatitude
 ) { }
